# api/swagger/swagger.yaml
basePath: /
definitions:
  application.CategoryRes:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  banner.Banner:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      index:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  common.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  domain.CartItem:
    properties:
      id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  infra.addToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  infra.cartResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.CartItem'
        type: array
      total:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  infra.updateCartItemRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
host: localhost:3000
info:
  contact:
    email: support@yadwy.com
    name: API Support
  description: This is the Yadwy backend service API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Yadwy Backend API
  version: "1.0"
paths:
  /banners:
    get:
      description: Get a list of all banners
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/banner.Banner'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Banners not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all banners
      tags:
      - banners
    post:
      consumes:
      - multipart/form-data
      description: Create a new banner with image upload (Admin only)
      parameters:
      - description: Banner name
        in: formData
        name: name
        required: true
        type: string
      - description: Banner display index
        in: formData
        name: index
        required: true
        type: integer
      - description: Banner image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/banner.Banner'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new banner
      tags:
      - banners
  /cart:
    delete:
      description: Remove all items from the cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart cleared
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Clear cart
      tags:
      - cart
    get:
      description: Get the current user's shopping cart with all items and total
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.cartResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the user's shopping cart
      parameters:
      - description: Product details to add to cart
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/infra.addToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Item added to cart
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - cart
  /cart/items/{productId}:
    delete:
      description: Remove a product from the cart
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Item removed
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Update the quantity of a product in the cart
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: Updated quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/infra.updateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item updated
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - cart
  /category:
    get:
      description: Get a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/application.CategoryRes'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - multipart/form-data
      description: Create a new category with image upload
      parameters:
      - description: Category name
        in: formData
        name: name
        required: true
        type: string
      - description: Category description
        in: formData
        name: description
        required: true
        type: string
      - description: Category image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"